refactored file system and pointed paths to newly updated file system
