+++ src/cssnhtml/account-creation-post-submit.php
@@ -5,7 +5,7 @@
 	</head>
 	<body>
 		<?php
-			require_once("sqlFunctions.php");
+			require_once("../php/sqlFunctions.php");
 
 			$conn = connectDB();
 			if(!$conn)
