+++ src/cssnhtml/password_reset_post_submit.php
@@ -1,5 +1,5 @@
 <?php
-    require_once("sqlFunctions.php");
+    require_once("../php/sqlFunctions.php");
 
     $conn = connectDB();
     if(!$conn)
